requirements.txt
fastapi
uvicorn
resemblyzer
numpy
scipy
psycopg2-binary
python-multipart